requirements.txt
asgiref==3.2.10
attrs==19.3.0
Django==3.0.8
djangorestframework==3.11.0
django-test-plus==1.4.0
factory-boy==2.12.0
Faker==4.1.1
iniconfig==1.0.0
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.0.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
